Changed how Module Numbering Works
Prior to this, all numbering was done manually. This was, in hindsight,
massively stupid.

The table of contents on the main page and the dropdown menu of modules
are now numbered lists, which makes the module numbering automatic and
less horrible.

Also removed many of the now-unnecessary Liquid filters which were being
used to strip the numbering from the titles of modules (WHY did I think
that was the best way to do things?)